namada_systems::parameters
Trait Read

pub trait Read<S> {
    // Required methods
    fn read(storage: &S) -> Result<Parameters>;
    fn masp_epoch_multiplier(storage: &S) -> Result<u64>;
    fn epoch_duration_parameter(storage: &S) -> Result<EpochDuration>;
    fn is_native_token_transferable(storage: &S) -> Result<bool>;
    fn epochs_per_year(storage: &S) -> Result<u64>;
    fn estimate_max_block_time_from_blocks_and_params(
        storage: &S,
        last_block_height: BlockHeight,
        num_blocks_to_read: u64,
    ) -> Result<DurationSecs>;
}

Abstract parameters storage read interface

Required Methods

fn read(storage: &S) -> Result<Parameters>

Read all parameters

fn masp_epoch_multiplier(storage: &S) -> Result<u64>

Read MASP epoch multiplier parameter

fn epoch_duration_parameter(storage: &S) -> Result<EpochDuration>

Read the the epoch duration parameter

fn is_native_token_transferable(storage: &S) -> Result<bool>

Read the is_native_token_transferable parameter

fn epochs_per_year(storage: &S) -> Result<u64>

Read the number of epochs per year parameter

fn estimate_max_block_time_from_blocks_and_params( storage: &S, last_block_height: BlockHeight, num_blocks_to_read: u64, ) -> Result<DurationSecs>

Return an estimate of the maximum time taken to decide a block, by sourcing block headers from up to num_blocks_to_read, and from chain parameters.

Object Safety

This trait is not object safe.