namada_systems
Module trans_token

Transparent token abstract interfaces

Structs

Amount

Amount in micro units. For different granularity another representation might be more appropriate.

DenominatedAmount

An amount with its denomination.

Denomination

Given a number represented as M*B^D, then M is the matissa, B is the base and D is the denomination, represented by this struct.

Enums

AmountError

AmountParseError

EventLevel

Indicates if an event is emitted do to an individual Tx or the nature of a finalized block

MaspDigitPos

The four possible u64 words in a Uint. Used for converting to MASP amounts.

UserAccount

A user account.

Constants

NATIVE_MAX_DECIMAL_PLACES

Maximum decimal places in a native token Amount and Change. For non-native (e.g. ERC20 tokens) one must read the denom_key storage key.

NATIVE_SCALE

Decimal scale of a native token Amount and Change. For non-native (e.g. ERC20 tokens) one must read the denom_key storage key.

Traits

Events

Abstract token events interface

Keys

Abstract token keys interface

Read

Abstract token storage read interface

Write

Abstract token storage write interface

Type Aliases

Change

A change in tokens amount

Result

Result of a storage API call.